ALL TESTS PASS, but the macro compiler is still unused.
Declaration mechanism is implemented but untested.  If a macro's argument is a declaration node, that node immediately enters scope.  When the end of the macro invocation is reached, all such declarations are popped out of scope.

It still does not deal with type expressions, which have to be evaluated (and parsed) in the outermost scope.  There must be another parsing instruction for this to be right.

I think primitive 355 is still not right.