ALL TESTS PASS.
Atoms now have properties (see primitives 122-125).  The property keys are also atoms, but the property values can be anything.

An object instance of AtomWithPropertiesDescriptor is automatically substituted for an AtomDescriptor object instance when the first property is added.  This ensures atoms only pay to store properties when that feature is actually used.

To do:  Use properties to implement object type naming.  Every time a type name is introduced, add information about the type and type name to a "system" property within each atom constituting the keyspace of the object type.  For example, a Point type with x and y keys would list the type Point and the name "Point" in a list within a special property of both of the atoms x and y.  A ThreeDPoint would add an entry for "ThreeDPoint" to the same lists within x and y, and also within z.