ALL TESTS PASS.
Object types names are no longer tracked by the AvailRuntime, but occur instead within a property of the atoms that are the keys of the very object types being named.

Primitives 68 and 69 were updated.  In particular, primitive 69 always succeeds now, and answers a set of applicable, locally-most-specific names.

Primitive 69 is not yet used, but it may be needed eventually for printing.