ALL TESTS PASS
Object types and Objects now print using the named type information that can be embedded in a special property of the constituent atoms.

Also started work on RestrictionSignatureDescriptor -- similar to a hybrid between a requires clause and returns clause, but without the need for a body.  Eventually requires and returns clauses will be replaced with a restrict clause, which will have another syntax when it's used in isolation.