ALL TESTS PASS.
The main goal here was to rearrange the Avail object hierarchy so that AvailObject could be declared final. All methods in the hierarchy are now final. I observed a performance boost of about 14% on my development machine, which was surprisingly high; I was really expecting Java's jitter to work a little harder at optimization, but perhaps the specifics of the problem were rather harder than I imagined. Perhaps I mismeasured somehow?

AbstractAvailObject has been introduced as the root of the Avail object hierarchy. It specifies the abstract capabilities of Avail objects, and provides some useful (final) methods.

AvailObjectRepresentation replaces AvailObjectUsingArrays, but sits in the middle of the lineage. It defines the concrete representation of an Avail object, and implements (as final) most of the abstract methods of AbstractAvailObject.

AvailObject is now the bottom of the Avail object hierarchy rather than the top. The delegation methods are implemented here, and someday should be machine generated. Descriptor dispatches are made directly using the "descriptor" field defined up in AbstractAvailObject (rather than through the descriptor() method, as previously).

Works in progress should (hopefully) only have to concern themselves with conflicts in AvailObject during integration.