ALL TESTS PASS
Completed the arithmetic operations (+, -, *, /, and comparisons) for floats, doubles, integers, and infinities.  Note that (double op anything) is a double, (float op anything but a double) is a float), etc.  The double and float infinities and NaNs are used as needed to satisfy these rules.

Introduced primitive type "number" above extended integers, floats, and doubles.

Renamed integerSlot(...)/objectSlot(...) family of operations to just slot(...) and setSlot(...).

Added simple tests of arithmetic.  Needs to be expanded.

Added test to verify assertions are enabled.

NOTE:  There is a known defect (discovered but not yet patched) in the code generator, whereby an outer variable is cleared prior to looping via restart of a continuation.  The variable's value is actually needed by the continuation when it restarts.