ALL TESTS PASS.
I resolved the twa module roots problem; see the lengthy comment for Tier-3/Tuple (below).

As a consequence, for maximum convenience, the AvailBuilderFrame should now be run with these VM arguments:

	-ea -DavailRoots=avail=${workspace_loc:Avail}/new-avail;old-avail=${workspace_loc:Avail}/avail
	
This will make the old and new libraries available to run.

Java:
com.avail.environment:
AvailBuilderFrame: Modules may now be double-clicked to run.

Avail (avail):
Kernel/Tier-3:
Tuple: Fixed a very insidious problem. The semantic restriction on "_[_]" implicitly depended on semantic restrictions on "_lower bound" and "_upper bound" defined in Tier-3/Integer. This module was not being imported (absent from "Uses" and "Extends"). When running the JUnit tests, they all passed ... because the AvailBuilder happened to linearize the "independent" modules Tier-3/Integer and Tier-3/Tuple such that the former was loaded first, causing the semantic restrictions on "_lower bound" and "_upper bound" to be present. When running AvailBuilderFrame with roots for both the old and new libraries defined, the linearization of these two modules occurred in the reverse order, causing compilation to fail. Tier-3/Tuple now imports Tier-3/Integer (which it should have all along), but problems like this are going to be common *and* difficult to resolve.