ALL TESTS PASS
Fixed a problem when an L2Chunk's numObjects() was being used incorrectly (or computed incorrectly), leading to failure in wipeObjectRegisters().

Split L2_SEND* instructions into separate L2_LOOKUP* and L2_INVOKE instructions.  The lookup is elided for monomorphic call sites.  The long term goals are (1) to inline the lookup decision tree logic directly into L2, and (2) to inline the invocations that look profitable.