ALL TESTS PASS
L2Operation is no longer an enumeration, but an abstract class whose subclasses are the specific operation types.  This reduces the size of file being operated on.  The operations are assigned ordinals in the order in which they are first needed by the optimizer.

It's unclear if primitives can eventually be reorganized in a similar way.  Perhaps a mechanism to dynamically load the primitives by synthesizing partially numeric class names.  The primary distinction from L2Operations is that Avail code must be able to refer to specific primitives somehow.

Moved some classes from com.avail.compiler to the new com.avail.builder.

builder.Repository.java is currently a commented-out sketch.