ALL TESTS PASS
Moved all primitives into separate files.  For example, the addition primitive was an enum value of Primitive with the name Prim1_Addition, but is now a subclass of Primitive with the name P_001_Addition.  They are in a dedicated package which is scanned for conformingly-named classes as needed.

Changed TypeConsistencyTest to use Primitive.Result as an example enum instead of Primitive, since the latter is no longer an enum.  Renamed AVAIL_PRIMITIVE_ENUM_POJO --> AVAIL_PRIMITIVE_RESULT_ENUM_POJO.