Here are some sample pages for the documentation, "Annals".
Main page is Annals.html. Other html pages are frame content. Bunnies and Pickles are present for effect only, and neither their content nor style are intended for real consideration. Only the Logic category link leads to the intended effect in the Methods frame, and the same for the _=_ method link in the methods frame (visible in Annals.html by default or after visiting the Logic link).