ALL TESTS PASS
Meaningful Avail stack frames are now output by P_256_EmergencyExit prior to the Java stack trace.  For each entry it shows the method name, module name, and line number.  And a Merry Xmas to all.

The same textual description is also available when examining an L2Interpreter in the Variables pane of the debugger (when Show Logical Structure is enabled and suitably configured).