ALL TESTS PASS
Highlight: COMPILED MODULES.  A cache of modules is maintained in a combination of RAM and a temporary file.  The first build takes as long as it used to, but subsequent builds of those same modules are wicked fast, completely bypassing parsing of the files.  Updating a module file will invalidate the cached representation for that module, causing it to be recompiled during the next build.  Unmodified successor modules will simply load on top.