ALL TESTS PASSED.
- Subtuple union ("∪_[_.._]", primitive 143) now accepts ∞ for the end
subscript.
- Corrected the semantic restriction for "_'s⁇default type".
- "Require:_[_.._]is ever valid" now allows a start index one greater
than the size of the tuple.
- Correct the semantic restriction for "_×_"; it produced an incorrect
type when one argument included an infinity and the other did not. (This
is still wrong in the old library, but I made a note there.)
- Defined and exported "_÷_" for two integer range types.
- Added a semantic restriction for "_-_" ([set, any]→set).
- Tuple summation ("Σ_") now supports extended integers; rewrote the
semantic restriction to support this widening.
- Added "quicksort_" and "quicksort_with_". Natural implementations of
"quicksort_" are provided for numbers and strings. (Natural ordering of
strings is character-wise code point, with length used to break prefix
ties.)