ALL TESTS PASS.
- Additional exception handling tests.
- I have determined that the problem with returning a value from an
exception handler is strictly limited to primitives 340 and 341. When
the chunk of code in AvailCodeGenerator that blesses conformant
functions with these methods is commented out, then it is possible to
answer a value from an exception handler. This is definitely a VM bug,
so I will write up an issue and assign it to Mark.