ALL TESTS PASs (except the macro test).
- Added "Network" to the Avail library. So far it only contains the
primitive bindings for server sockets, client sockets, IP addresses, and
socket options, as well as a handful of unit tests.
- Added primitives for getting the local address of a server socket and
the local and remote peer of a client socket.
- Fixed numerous bugs in the socket primitives.
- Renumbered some of the socket primitives.
- Implemented several missing methods in TupleDescriptor relating to
interoperation with ByteBufferTupleDescriptor.
- Fixed systematic bitmask errors in ByteArrayTupleDescriptor and
ByteBufferTupleDescriptor.
- Removed o_IsByteTuple from TupleDescriptor. General tuples now answer
false, even if they happen to be tuples of only bytes. The test is
really supposed to indicate whether the "raw byte" family of methods is
available.
- Fixed several bugs in monitors related to awaiting/signaling
conditions.
- Fixed a bug in "Await_" [semaphore] that prevented a permit from being
consumed.
- Added missing semantic restrictions for the atomic this-and-that
synchronization primitives. I also added unsafe versions of same (for
generic use).