Introduced the Lincolnshire Poacher example.
All Java tests pass.
All Avail tests pass.

/avail/Avail/Foundation:
Synchronization: Chucked some code that was explicitly tracking the
nesting of reentrant mutexes. This was never necessary or useful anyway,
plus I discovered a bug in it. It was better to discard the tracking
altogether than fix the bug.

/examples:
Examples: New module. It acts like a "representative" of the root, and
makes it easy to build all of the examples, i.e., just build this one
instead of each example separately.
Lincolnshire Poacher: Modernized the synchronous numbers station
introduced in edition #3 of the Avail newsletter. For the first time, it
is now available as a running example. The entry points are

	- "Start Lincolnshire Poacher", which starts the server.
	- "Stop Lincolnshire Poacher", which stops the server. (This is
automatically called when the module is unloaded.)
	- "Eavesdrop", which listens to a single mysterious "message".