Resolved issue #81. It is now possible to prompt the user to disambiguate an ambiguous command.
All Java tests pass.
All Avail tests pass.

[VM]
AvailBuilder: Introduced inner class CompiledCommand. Refactored
attemptCommand() to accept an additional argument: a Continuation2 that
itself accepts a List of CompiledCommands (corresponding to the possible
parses of a command) and a Continuation1 (that accepts the selected
parse, or null, and proceeds).

[workbench]
SubmitInputAction: A JOptionPane is now displayed in the event of an
ambiguous command parse. The pane contains all valid parses. The user
may select the intended parse, or cancel.