Introducing… the Avail server!
An Avail server [com.avail.server.AvailServer] hosts an Avail runtime
for concurrent use by multiple clients. The connection protocols include
"basic" (4-byte size-prefixed UTF-8 strings) and, much more exciting,
WebSocket. (Support is present for binary WebSocket, but only text is
currently utilized.) A client interacts with the server by sending
textual commands and receiving JSON responses.

The functionality of the server is currently rather limited, but the
server can already build modules and run entry points. From an Eclipse
environment, the server can be launched with the following VM options:

-ea -Xdock:name="Avail Server"
-Xdock:icon=${workspace_loc:Avail}/distro/images/AvailHammer.png
-Djava.util.logging.config.file=${workspace_loc:Avail}/distro/etc/logging.properties
-DavailDeveloper=true
-DavailRoots=avail=${workspace_loc:Avail}/repositories/avail.repo,${workspace_loc:Avail}/distro/src/avail;examples=${workspace_loc:Avail}/repositories/examples.repo,${workspace_loc:Avail}/distro/src/examples;website=${workspace_loc:Avail}/repositories/website.repo,${workspace_loc:Avail}/distro/src/website

A very raw web client [distro/ide/TestClient.html] is provided for
interacting with a running server. Type "commands" and press the "Send"
button to receive a complete list of commands.

For example, the following script will start a game of Wump the Wumpus:

	build module /examples/Wump the Wumpus
	run Play Wump the Wumpus

This is groundwork for a web-based Avail IDE. The AvailBuilderFrame will
continue to operate for the time being, but will eventually be abandoned
in favor of the web-based IDE.