Avail values can now render themselves as JSON ("_→JSON"). This was a significant amount of work, so probably some of the paths are buggy. We'll get these issues hammered out as time goes on, and as we beginning building real web tools for Avail.
The test client is able to render (prettily) JSON results of Avail
commands (entry point expressions). If a result cannot be understood as
JSON, then it is rendered linerally and in blue instead.

The test client moved from /distro/ide to /distro/web/ide, in
anticipation of more content in web (like custom clients for existing
and future Avail example programs).

I also fixed a handful of bugs.