The Ship of Stories got a major facelift today. I built a custom web client for it. The web client should be sufficient for any adventure built with Avail's Choosable Path DSL, but it's currently hard coded for The Ship of Stories. (I'll probably get around to fixing that soon.)
I've tested the web client on modern versions of Safari and Chrome… and
nowhere else. The behavior on the two browsers was virtually identical,
so I'm happy. Chrome's treatment of the Escape key seems a little bit
wonky, and I'm not sure that it's working in conjunction with the Shift
modifier. I'll look into that more later.

The web client is here:

	/distro/web/clients/The Ship of Stories.html

Just point your web browser at it using a file URL if you want to give
it a try.

To play, just click on the pink boxes (with italic text) that represent
the possible choices. You can also hit the numbers 1 through 9 for the
first 9 choices presented by a scene; the 10th choice must be selected
by the mouse, but there are no scenes in The Ship of Stories with this
many choices, so the limitation is moot.

You can restart the game at any time by pressing Space. You can quit the
game at any time by pressing Escape. Should you lose the adventure, this
information will be presented to you explicitly in a red box at the
bottom of the window.

This was a lot of fun for me. I hope that you will enjoy it for a few
minutes too. :)