Class definition now permits specification of the accessor, mutator, and referrer names for fields.
Immutable fields permit specification of the accessor only, in a
parenthetical that follows the field declaration (and precedes the comma
before the next field, if any). The optional keyword "accessor" is
permitted before the custom accessor name (for symmetry with mutable
field naming).

Mutable fields permit specification of each, in a parenthetical that
follows the field declaration (and precedes the comma before the next
field, if any). One of the keywords "accessor", "mutator", and
"referrer" must (uniquely) precede the custom name. Custom name
specifications are comma-separated.

For example, from a personal project of mine:

Public class "score" extends object
	with field
		public base model element : model element ("_'s⁇basis")
	with mutable field
		public score modifiers : <score modifier…|>
			(accessor "_'s⁇modifiers", mutator "_'s⁇modifiers:=_");

The idea here is to generate "_'s⁇basis" instead of "_'s⁇base model
element", and "_'s⁇modifiers" / "_'s⁇modifiers:=_" instead of "_'s⁇score
modifiers" / "_'s⁇score modifiers:=_".