ALL TESTS PASSED.  Added the macro "For each…of_§do_" for set.  This macro is a set-iterating loop that invokes the action block once for every element in the set.  The macro determines the strongest type for the elements of the set and assigns that type to the input of the function block.
Added the macro "For each…,…of_§do_".  This macro is a tuple-iterating
loop that invokes the action block once for every element in the tuple.
The index of the element is made available in the scope of the function.
The macro determines the strongest type for the elements of the tuple
and assigns that type to the input of the function block.