ALL TESTS PASSED.  Restructured the entirety of the Avail testing directory.  Created a new directory, Tests.avail, at the same level as Foundation.  Moved Foundation Tests.avail into the new directory.  In Tests.avail, created Data Abstractions Tests.avail.  Moved all data abstraction tests out of Data Abstractions.avail (most lived at the end of modules they were testing) into this new directory.
Added the macro "For each…of_§do_".  This macro is a stream-iterating
loop that invokes the action block once for every element in the stream.
The macro determines the strongest type for the elements of the stream
and assigns that type to the input of the function block.  Created
Stream Tests.avail and added tests for this functionality.

Added the macro "For each…→…of_§do_".  This macro is a map-iterating
loop that invokes the action block once for every key→value pair in the
map.  The macro determines the strongest type for the key and the value
of the map and assigns those types to the input of the function block.
Added tests for this functionality to Map Tests.avail.