ALL TESTS PASS.
[VM]
AbstractAvailCompiler: Top- and bottom-valued expressions are now
correctly rejected when they are the leading arguments of macros (and
the argument is not suffixed with the ! metacharacter for phrase
wrapping).
P_384_SendMethod: Now correctly returns the method instead of the
bundle.

[Avail]
/avail/Avail/Foundation/Bootstrap:
Infallible Primitives: Removed "_+_" (Primitive 703) for tuple
appending. It is largely subsumed by the "_++_" macro (see below). I
reintroduced it as "_with_appended" into "Early Tuples" in Foundation. I
am going to introduce a concatenating assignment operator in that
module, as a macro, and have the macro generate calls of
"_with_appended". This will provide efficient accumulation into tuple
variables, a pattern common in the Avail library. (Have no fear,
Robbert, uses of your beautiful tuple append work will be back very
soon!)

/avail/Avail/Foundation:
Early Tuples: Removed "«_‡++»". Introduced the macro "_++_" as a
replacement. It produces either literal tuples or sends of
"concatenate_", as appropriate. There are numerous implementations of
"_++_" to maximize folding and compaction of concatenation.

/avail/Tests/Foundation Tests:
Tuple Tests: Updated "tuple appending" because of changes related to
tuple concatenation. When I commit the new compound concatenating
assignment operator, I will rewrite this test to be based around it.