ALL TESTS PASS.
[VM]
AvailRuntime, Interpreter, Primitive, L2Translator,
P_076_SetResultDisagreedWithExpectedTypeFunction: Changed the signature
of resultDisagreedWithExpectedTypeFunction to <function, type, ↑any>.
P_219_FunctionMessage: New primitive ("_'s⁇message") to extract the name
of the enclosing method from a function.

[Avail]
/avail/Avail/Foundation:
Exceptions: A result-disagreed-with-expected-type exception now reports
the message that failed, the expected type, and the actual value
generated at runtime. This is a huge help when debugging faulty semantic
restrictions.