ALL TESTS PASS.
[Avail]
/avail/Avail/Foundation:
Exceptions: "Guard|guard_«intercept_»«ensure_»" did not call the ensured
function in the event that intercepts existed but not for the raised
exception. This has been fixed. (Thanks for noticing this, Rich! This
defect was apparently always present, and not a consequence of my work
earlier today.)

/avail/Tests/Foundation Tests:
Exception Tests: Added a regression test for the scenario described
above.