A basic Code Coverage Report is now in service.
This report displays selected details about the functions represented by
CompiledCode objects that have a non-Nil module field. (CompiledCode
objects with Nil for a module were constructed by the VM for
bootstrapping.) The preexisting VM report Actions (ReportAction and
ClearReportAction) were renamed to be more specific about what they were
reporting.

The controls to display these reports and reset their data are in the
Avail Workbench’s Developer menu, which only displays if you include the
following in your VM args:

-DavailDeveloper=true

At a future time, the Code Coverage Report category may grow to include
smaller, easier-to-read reports that are, for example, restricted by
module name or restricted to include only data on which functions did or
did not run.