WIP - tests still run except MessageSplitterTest (stubbed).
Calls of bundles with permuted arguments still don't compile.  Parsing is still much slower than before the introduction of argument type filtering.

There's still an old bug in the L2 translator, which shows up when translating “_” operation for timestamp.  Register 14 should be typed as exploded timestamp, but it's just timestamp.  Reproduced by running this in the Availuator:

    !“now” ++ “now” ++ “now” ++ “now” ++ “now” ++ “now” ++ “now” ++ “now” ++ “now” ++ “now” ++ “now”