WIP - tests still run except MessageSplitterTest (stubbed).
Calls of bundles with permuted arguments now compile (and execute) correctly.

Fixed bug in tree tuples – it was allowing empty subtuples, so when it looked up the subtuple that started at the needed offset, it tried to access the 1st element of the (empty) subtuple and failed.  Added an additional consistency check to ensure subtuples can't be empty.