Delayed APPEND_ARGUMENT by introducing WRAP_IN_LIST parsing instruction.  This makes the initial few parsing instructions of most plans more uniform, and a bit more efficient to parse in aggregate.
Note 1:  Methods starting with Optional expressions happen to attempt PUSH_FALSE in parallel with dispatch along the positive case – typically a PARSE_PART, which means it just occupies another entry of the token dispatching map.  We can improve this in the future by code-splitting leading-Optional methods, postponing merge of control flow by duplicating perhaps one more expression in the root sequence, allowing the PUSH_FALSE to happen after some other parsing.  We may need to introduce a SWAP parsing operation for the case where the next thing to be parsed might be an argument.

Note 2:  It might be beneficial to categorize all plans as leading-part, leading-argument, or potentially both.  Keep two root bundle trees in the loader.  Use the leading-part bundle tree first when parsing a subexpression, then apply the leading-argument bundle tree repeatedly.  The existing mechanism of passing a pre-parsed argument or null will still be needed.

Unrolled initial iterations of Groups, even when the list type was homogenous, based on the minimum list size in the type signature.

Fixed some example modules with entry points that didn't run because of the recently introduced "…$" and "…#".