Improved fast-loading performance.
Added P_AtomicAddToMap ("_[_]:=_") and P_KeyInVariableMap ("_∈_").  A few key places were reading a map from a global variable, adding a key/value pair, and writing it back to the variable, which disabled the fast-loading optimization.  Using the new primitives allows the fast-loading to stay enabled (using the new LoadingEffectToAddToMap helper).

Tweaked deserializer for performance, creating a reusable operand array.

Small tweak to ParsingOperation to flag whether an operation can be usefully performed when there is a pre-parsed first argument that has not yet been consumed.  If avoids queueing a work unit in that circumstance.