A series of Optionals like "Public?enemy?number one" is now transformed into code splitting as deep as necessary to ensure a token or argument is reached prior to having to push true or false for any of the options.  The arguments are later reordered as needed on the stack.  This eliminates push(false) from the root bundle tree.
Primitives that add a definition or restriction or seal now fail with E_CANNOT_DEFINE_DURING_COMPILATION if the loader is not currently running a top-level statement.  This prevents type declarations and other compile-time code from altering the bundle trees, eliminating mutual exclusion locks when checking whether a bundle tree needs to be expanded.

Added volatileSlot() and setVolatileSlot() to support the weaker locking of the bundle tree's UNCLASSIFIED map.