Rebuilt template expansion.
Ctrl+Space now attempts "smart completion" of a template. Any text from the caret to the previous word boundary is used as a prefix to filter the available templates. If the list contains no items, then nothing happens. If the list contains only one item, then the prefix is immediately replaced with the template expansion. If the list contains multiple items, then a context menu is displayed at the caret which includes the matching items; upon selection of a menu item, the prefix is replaced with the result of expanding the selected template. These changes bring behavior in line for modern IDEs.