First pass of JIT L2-to-JVM translation.
In order to make this happen relatively quickly, we truncated our ambitions considerably. Since L2 is implemented as a reified instruction set, we mostly just generate calls to the runAction() method of the L2Instructions. There are a few cases, such as L2_JUMP*, where we actually generate "legitimate" JVM translations. This is cheesy, but was intended to enable fast, accurate implementation that separated the translation framework from individual instruction translation.

The intention, of course, is to effect a fully "legitimate" translation of optimized L2 chunks to JVM. This is considerably more work, but should still be achievable relatively quickly.

Even the cheesy implementation produces a 2x speed gain, so yay.