Added floating point number lexer
A valid floating point number must have a point, must have 1+ digits in
the integer position, must have 1+ digits in the decimal position, may
use 'e' to indicate an exponent (in which case it must have 1+ digits in
the exponent position), may have an 'f' at the end to indicate that
float should be used instead of double, and may have an underscore
character to separate groups of 1+ digits in any of the three digit
positions.
- Added a lexer to recognize this as a lexeme and adjusted everything to
  use it. This required some reordering/relocations of definitions,
  imports, and exports.
- Also added tests to recognize underscorey numbers as valid
  floats/doubles.

All tests pass ONLY when L2 is turned off by upping L2Chunk.java
countdownForNewCode retval up to 1000000000. When set down to 1, fails
with "Unrecoverable problem during optimization".