Let availc handle its own options
The Avail compiler command can handle its own options/flags, so for now
just skip the getopts call and let it parse them for itself. This isn't
really a loss of functionality since there are roots-manipulating
arguments among them.

Also move the usage help flag to -? to match the compiler.