Fix semantic restriction on tuple element removal
Typo in disjointedness check was comparing against the leading types of
`tuple` itself, rather than the argument. This lead to an incorrect
rejection of parse when the first leading type of a tuple was disjoint
from the type of the element being removed.  (e.g. `<1, 2, 3> - 2`
would fail)

Also refactored check to use `each of_satisfies_` to make logic more
explicit.