Add drain tests
Also, a number of the streams tests were testing iterators instead of
streams, so convert them to the appropriate. In general the coverage of
the streams test suite is pretty bad. I added a select few low-hanging
fruit as well, but if I keep expanding the scope of this PR it will never
be done.