Fixed structure view bug when closing an editor
Previously, if 2+ editors were open and one was closed, the structure view would update the title and path but NOT the manifest entries; it would be stuck showing the entries for the editor that was just closed. StructureViewPanel updateView() should not be used when it is expected that the editor in focus has changed, as it does not fetch the editor's manifest entries and relies on them being passed.